Template and Profile update

+++ services/backend/app/crud.py
@@ -8,7 +8,7 @@ from app.exceptions import Exc
 from app.security import get_password_hash
 
 
-def create_generic(db: Session, model, object_in, profile_id: str):
+def create_generic_profile_child(db: Session, model, object_in, profile_id: str):
     model_obj = model(**object_in.dict())
     db.add(model_obj)
     model_obj.profile_id = profile_id
@@ -17,6 +17,8 @@ def create_generic(db: Session, model, object_in, profile_id: str):
 
 def delete_generic(db: Session, model, obj_id: str):
     model_obj = db.query(model).filter(model.id == obj_id).first()
+    if model_obj is None:
+        raise Exc.ObjectNotFoundException("Object")
     db.delete(model_obj)
     db.commit()
 
@@ -34,7 +36,7 @@ def get_user_by_username(db: Session, username: str):
 def create_user(db: Session, user_in: schemas.UserCreate):
     user_inDB = get_user_by_username(db, user_in.username)
     if user_inDB is not None:
-        raise Exc.UsernameTakenException
+        raise Exc.NameTakenException("Username")
     password_hash = get_password_hash(user_in.password)
     user_obj = models.User(
         username=user_in.username,
@@ -49,7 +51,7 @@ def create_user(db: Session, user_in: schemas.UserCreate):
 def update_user(db: Session, user_in: schemas.UserUpdate, user_id: str):
     user_obj = get_user_by_id(db, user_id)
     if user_obj is None:
-        raise Exc.UserNotFoundException
+        raise Exc.ObjectNotFoundException("User")
     user_obj.username = user_in.username
     user_obj.password = get_password_hash(user_in.password)
     db.commit()
@@ -59,6 +61,26 @@ def update_user(db: Session, user_in: schemas.UserUpdate, user_id: str):
 
 def delete_user(db: Session, user_id: str):
     delete_generic(db, models.User, user_id)
+
+
+def promote_user(db: Session, user_id: str):
+    user_obj = get_user_by_id(db, user_id)
+    if user_obj is None:
+        raise Exc.ObjectNotFoundException("User")
+    user_obj.is_admin = True
+    db.commit()
+    db.refresh(user_obj)
+    return user_obj
+
+
+def upgrade_user_to_premium(db: Session, user_id: str):
+    user_obj = get_user_by_id(db, user_id)
+    if user_obj is None:
+        raise Exc.ObjectNotFoundException("User")
+    user_obj.is_premium = True
+    db.commit()
+    db.refresh(user_obj)
+    return user_obj
 
 
 def get_user_profile_by_public_name(db: Session, public_name: str):
@@ -73,7 +95,7 @@ def get_user_profile_by_id(db: Session, id: str):
 def validate_publicname(db: Session, public_name: str):
     user_profile_inDB = get_user_profile_by_public_name(db, public_name)
     if user_profile_inDB is not None:
-        raise Exc.PublicNameTakenException
+        raise Exc.NameTakenException("Public name")
 
 
 def create_user_profile(
@@ -130,7 +152,7 @@ def update_each_or_create(
                 old_object_inDB = db.query(model).filter(
                     model.id == obj_in.id).update(obj_in.dict())
             else:
-                create_generic(db, model, obj_in, profile_id)
+                create_generic_profile_child(db, model, obj_in, profile_id)
 
 
 def filter_out_removed(db: Session, model, incoming_objects: List):
@@ -142,7 +164,7 @@ def filter_out_removed(db: Session, model, incoming_objects: List):
 def update_user_profile(db: Session, profile_in: schemas.UserProfileUpdate, user_id: str):
     profile_inDB = get_user_profile_by_id(db, profile_in.id)
     if profile_inDB is None:
-        raise Exc.ProfileNotFoundException
+        raise Exc.ObjectNotFoundException("User Profile")
     if profile_inDB.public_name != profile_in.public_name:
         validate_publicname(db, profile_in.public_name)
     
@@ -175,3 +197,48 @@ def update_user_profile(db: Session, profile_in: schemas.UserProfileUpdate, user
 
 def delete_user_profile(db: Session, profile_id: str):
     delete_generic(db, models.UserProfile, profile_id)
+
+
+def get_templates(db: Session, user: models.User):
+    if user.is_premium:
+        return db.query(models.Template).all()
+    return db.query(models.Template)\
+        .filter(models.Template.premium == False).all()
+
+
+def get_template_by_id(db: Session, id: str):
+    return db.query(models.Template)\
+        .filter(models.Template.id == id).first()
+
+
+def get_template_by_name(db: Session, name: str):
+    return db.query(models.Template)\
+        .filter(models.Template.name == name).first()
+
+
+def create_template(db: Session, template_in: schemas.TemplateCreate):
+    existing_template = get_template_by_name(db, template_in.name)
+    if existing_template is not None:
+        raise Exc.NameTakenException("Template name")
+    template_obj = models.Template(**template_in.dict())
+    db.add(template_obj)
+    db.commit()
+    db.refresh(template_obj)
+    return template_obj
+
+
+def update_template(db: Session, template_in: schemas.TemplateUpdate, template_id: str):
+    template_obj = get_template_by_id(db, template_id)
+    if template_obj is None:
+        raise Exc.ObjectNotFoundException("Template")
+    
+    template_obj.name = template_in.name
+    template_obj.content = template_in.content
+    template_obj.premium = template_in.premium
+    db.commit()
+    db.refresh(template_obj)
+    return template_obj
+
+
+def delete_template(db: Session, template_id: str):
+    delete_generic(db, models.Template, template_id)
